I authenticated Locations and Industries
I authenticated the locations and industries so everyone can see them, like when searching or creating jobs and internships. But only Admin can add them or delete them. Also nonADMINs cannot see the dashboard if they write the URL it shows a text telling them that the admin is the only one who has access. And shows a link to the login page.